Refactor pybullet util functions into backend.pybullet
See discussions in #61
